Correct `config.exs` example
At least in the current implementation, Sugar requires the `:https` configuration key to be either correctly configured (with SSL key/cert and port) or explicitly set to `false`.  This change addresses that and fixes part of Sugar's Issue #61 by providing the necessary instruction for a new user to be able to run `mix server` without strange Cowboy errors.