Fix controller testing docs
The test example was calling router functions on the controller; a test written to emulate the example would not compile.

This new example actually works the controller via the router, so it's not exactly a unit test, but it will have the desired effect. A better example would probably call the controller's :index action directly.